Update library5.py
I added the code for my function and its test here. The function requires 3 things: 
1. A list of audio arrays(audio_arrays)
2. A sample rate which should be an integer(sample_rate)
3. The fade percentage which is from 0 to 50.(fade_percentage) An error will be raised if the percentage is not within that limit. (and a default of 15 has been set)
Th function adds a fade in and fade out effect to each of the audios in the list of arrays. 
The function returns a list of smoothed arrays.

About the test:
It checks three things:
1. It tests to see that the length of each audio has not been changed. 
2. that the fade-in has been done well.
3. That the fade-out has been done well.

This code was double-checked by Chatgpt 4.