Update library3
I added the mute_segments function to library3.
It gets the list of audio segments and the indices of the segments that we want to mute, as inputs and replace those segments with zeros.
The function returns the list of the audio with specified muted segments and unchanged other segments.
The TestMuteSegments is also added.